+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the Aptos TypeScript SDK will be captured in this file. This changelog is written by hand for now. It adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+# Unreleased
+- Add `is_enum` field to `MoveStruct`.
+
 # 5.2.0 (2025-12-10)
 
 - Add `http2` as an optional parameter in `ClientConfig`